client/requirements.txt
argparse==1.2.1
astroid==1.3.6
Babel==1.3
coverage==3.7.1
docopt==0.6.2
eno==0.0.17
envoy==0.0.3
flup==1.0.3.dev-20110405
humanize==0.5.1
itsdangerous==0.22
logilab-common==0.63.2
meld3==1.0.2
mock==1.0.1
netifaces==0.8
nose==1.3.4
openbts==0.1.7
Paste==1.7.5.1
phonenumbers==6.2.0
psutil==3.1.1
psycopg2==2.6
pylint==1.4.3
PyYAML==3.10
pyzmq==14.5.0
requests==2.6.2
six==1.9.0
sms-utilities==0.0.3
snowflake==0.0.2
web.py==0.37
wsgiref==0.1.2
Twisted==15.1.0
python-dateutil==2.4.2
pytz==2015.7
enum34==1.0.4
